Fixed pytest failing for a default build
glob_sort_captured would fail if a single cfg file was passed to it. A list within the function only fills if the file contains "[]" or "*", and so an assertion that the list be of size==1 would fail. The condition to handle a single cfg file was after the assert line, and so was moved to the beginning of the function to force an early "return", and pytest now completes successfully

Other minor edits
 - source.organizeImports set to "explicit" not "true"
    as part of newer VSCode version
 - Added the IPython dependency to alleviate a
    failed pytest