Allowed float handling of file sorting
Added two new abilities to file sorting. Files with floats for captures ("[]") are now handled, and if a capture has multiple floats, then the last instance is used for sorting.

Also fixed a minor bug caused by using an "or" instead of "and" which didn't allow for a single capture type to be used